Rename rulesmarketplace.js to marketplace.js
